recipe-auth/requirements/prod.txt
Django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-oauth-toolkit==1.2.0
